@@ -18,18 +18,20 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.3
+  collection: ^1.15.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  json_serializable: ^3.3.0
-  build_runner: ^1.10.0
+  json_serializable: ^4.1.4
+  build_runner: ^2.1.2
+  http: ^0.13.3
 
 flutter:
   uses-material-design: true
