@@ -1,20 +1,98 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
+import 'package:hotels/models/hotel.dart';
+import 'package:http/http.dart' as http;
 
 class HomeView extends StatefulWidget {
-  HomeView({Key key}) : super(key: key);
+  HomeView({Key? key}) : super(key: key);
 
   @override
   _HomeViewState createState() => _HomeViewState();
 }
 
 class _HomeViewState extends State<HomeView> {
+  bool isLoading = false;
+  List<Hotel> hotels = [];
+
+  @override
+  void initState() {
+    super.initState();
+    getData();
+  }
+
+  getData() async {
+    setState(() {
+      isLoading = true;
+    });
+    final response = await http.get(Uri.parse(
+        'https://run.mocky.io/v3/ac888dc5-d193-4700-b12c-abb43e289301'));
+    var data = json.decode(response.body) as List;
+    hotels.addAll(data.map<Hotel>((hotel) => Hotel.fromJson(hotel)));
+    setState(() {
+      isLoading = false;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(),
-      body: Center(
-        child: Text("Hotels"),
+      appBar: AppBar(
+        actions: [
+          IconButton(onPressed: () {}, icon: Icon(Icons.list)),
+          IconButton(onPressed: () {}, icon: Icon(Icons.grid_view)),
+        ],
       ),
+      body: isLoading
+          ? Center(child: CircularProgressIndicator())
+          : ListView(
+              children: <Widget>[
+                ...hotels.map((hotel) {
+                  return Card(
+                    elevation: 2.0,
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(10.0),
+                    ),
+                    child: Column(
+                      children: <Widget>[
+                        Container(
+                          child: ClipRRect(
+                            borderRadius: BorderRadius.only(
+                              topLeft: Radius.circular(10.0),
+                              topRight: Radius.circular(10.0),
+                            ),
+                            child: Image(
+                              image:
+                                  AssetImage('assets/images/${hotel.poster}'),
+                              fit: BoxFit.fitWidth,
+                            ),
+                          ),
+                        ),
+                        Container(
+                          padding: EdgeInsets.only(left: 10, right: 10),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: <Widget>[
+                              Text(hotel.name),
+                              ElevatedButton(
+                                onPressed: () {
+                                  Navigator.pushNamed(
+                                    context,
+                                    '/detailsInfo',
+                                    arguments: hotel.uuid,
+                                  );
+                                },
+                                child: Text('Подробнее'),
+                              )
+                            ],
+                          ),
+                        ),
+                      ],
+                    ),
+                  );
+                }).toList(),
+              ],
+            ),
     );
   }
 }
